Created folders, deleted comments
