+++ Albums screen of the Photos app/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
