+++ Albums screen of the Photos app/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
