+++ Albums screen of the Photos app/Resources/Assets.xcassets/Contents.json
