+++ Albums screen of the Photos app/Resources/Info.plist
