+++ Albums screen of the Photos app/Sources/App/AppDelegate.swift
@@ -1,0 +1,16 @@
+//
+//  AppDelegate.swift
+//  Albums screen of the Photos app
+//
+//  Created by Kristina Korotkova on 12/01/23.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        return true
+    }
+}
