+++ Albums screen of the Photos app/Sources/App/SceneDelegate.swift
@@ -1,0 +1,17 @@
+//
+//  SceneDelegate.swift
+//  Albums screen of the Photos app
+//
+//  Created by Kristina Korotkova on 12/01/23.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let _ = (scene as? UIWindowScene) else { return }
+    }
+}
