--- Albums screen of the Photos app/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  Albums screen of the Photos app
-//
-//  Created by Kristina Korotkova on 12/01/23.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
